Split css/js files
