+++ simple-chat/src/chat/chat.css
@@ -1,0 +1,47 @@
+.messages {
+    flex-grow: 1;
+    background-color: #f5f5f5;  
+    padding: 80px 10px 60px 10px;
+    display: flex;
+    flex-direction: column;
+    justify-content: flex-end;
+}
+
+.form {
+    position: fixed;
+    bottom: 0;
+    width: inherit;
+    z-index: 1;
+
+    display: inline-block;
+    border: 1px solid rgba(25, 25, 25, 0.32);
+    display: flex;
+    height: 50px;
+    align-items: center;
+    padding: 0 10px;
+    background-color: #fff;
+}
+
+.form__attachment {
+    color: #707070;
+}
+
+form {
+    flex-grow: 1;
+}
+
+input[type=submit] {
+    visibility: collapse;
+}
+
+input::placeholder {
+    color: #a1a1a1;
+}
+
+input {
+    font-size: 16px;
+    padding: 10px;
+    width: 100%;
+    border: 0;
+    outline: none;
+}
