+++ simple-chat/src/index/index.css
@@ -1,0 +1,7 @@
+.chats {
+    flex-grow: 1;
+    background-color: #fff;  
+    padding: 80px 10px 20px 10px;
+    display: flex;
+    flex-direction: column;
+}
