+++ simple-chat/src/index/index.js
@@ -1,0 +1,7 @@
+import './index.css';
+import '../styles/global.css';
+import '../styles/reset.css';
+import '../styles/header.css';
+import '../styles/float-button.css';
+import '../styles/chat-item.css';
+import '../styles/status.css';
