+++ simple-chat/src/scripts/scrollToTheBottom.js
@@ -1,0 +1,3 @@
+export const scrollToTheBottom = () => {
+    window.scrollTo({top: document.body.scrollHeight});
+}
