+++ simple-chat/src/styles/message.css
@@ -1,0 +1,29 @@
+.message {
+    background-color: #f3e5f5;
+    display: flex;
+    flex-direction: column;
+    margin-bottom: 10px;
+    padding: 10px 15px;
+    max-width: 85%;
+    align-self: flex-end;
+    border-radius: 20px;
+    word-wrap: break-word;
+}
+
+.message__bottom {
+    display: flex;
+    justify-content: flex-end;
+}
+
+.message__text {
+    color: #1a181a;
+    font-size: 16px;
+}
+
+.message__time {
+    color: #706a71;
+}
+
+.message__status {
+    margin-left: 5px;
+}
