+++ simple-chat/src/styles/status.css
@@ -1,0 +1,4 @@
+.status {
+    font-size: 18px;
+    color: #9837b0;
+}
