+++ simple-chat/webpack.config.js
@@ -11,11 +11,12 @@ const BUILD_PATH = path.resolve(__dirname, 'build');
 module.exports = {
     context: SRC_PATH,
     entry: {
-        index: './index.js',
+        index: './index/index.js',
+        chat: './chat/chat.js',
     },
     output: {
         path: BUILD_PATH,
-        filename: 'bundle.js',
+        filename: '[name].bundle.js',
     },
     module: {
         strictExportPresence: true,
@@ -33,16 +34,7 @@ module.exports = {
                 ],
             },
             {
-                test: /shadow\.css$/,
-                include: SRC_PATH,
-                use: [
-                    {
-                        loader: 'css-loader',
-                    },
-                ],
-            },
-            {
-                test: /index\.css$/,
+                test: /\.css$/,
                 include: SRC_PATH,
                 use: [
                     {
@@ -57,11 +49,17 @@ module.exports = {
     },
     plugins: [
         new MiniCSSExtractPlugin({
-            filename: 'style.css',
+            filename: '[name].style.css',
         }),
         new HTMLWebpackPlugin({
             filename: 'index.html',
-            template: './index.html',
+            template: './index/index.html',
+            chunks: ['index'],
+        }),
+        new HTMLWebpackPlugin({
+            filename: 'chat/chat.html',
+            template: './chat/chat.html',
+            chunks: ['chat'],
         }),
     ],
 };
